+++ examples/factory_firmware/main/task_flow_module/tf_module_debug.c
@@ -13,7 +13,7 @@ static const char *TAG = "tfm.debug";
 static void __event_handler(void *handler_args, esp_event_base_t base, int32_t id, void *p_event_data)
 {
     tf_module_debug_t *p_module_ins = (tf_module_debug_t *)handler_args;
-    uint8_t type = ((uint8_t *)p_event_data)[0];
+    uint32_t type = ((uint32_t *)p_event_data)[0];
     switch (type)
     {
         case TF_DATA_TYPE_BUFFER: {
