+++ examples/factory_firmware/main/task_flow_module/tf_module_sensecraft_alarm.c
@@ -48,7 +48,7 @@ static void __event_handler(void *handler_args, esp_event_base_t base, int32_t i
     struct tf_module_sensecraft_alarm_params *p_params = &p_module_ins->params;
     ESP_LOGI(TAG, "Input shutter");
 
-    uint8_t type = ((uint8_t *)p_event_data)[0];
+    uint32_t type = ((uint32_t *)p_event_data)[0];
     if( type !=  TF_DATA_TYPE_DUALIMAGE_WITH_AUDIO_TEXT) {
         ESP_LOGW(TAG, "unsupport type %d", type);
         tf_data_free(p_event_data);
